pyproject.toml
[project]
name = "school-managment-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "cffi==2.0.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "fastapi==0.129.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "idna==3.11",
    "psycopg2==2.9.11",
    "psycopg2-binary>=2.9.11",
    "pwdlib==0.3.0",
    "pycparser==3.0",
    "pydantic-core==2.41.5",
    "pydantic[email]==2.12.5",
    "pyjwt==2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy==2.0.46",
    "starlette==0.52.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn[standard]==0.41.0",
]
